requirements.txt
anthropic
python-dotenv
streamlit
pymongo
